+++ src/hotspot/cpu/s390/sharedRuntime_s390.cpp
@@ -2056,7 +2056,7 @@ nmethod *SharedRuntime::generate_native_wrapper(MacroAssembler *masm,
   Label no_reguard;
 
   __ z_cli(Address(Z_thread, JavaThread::stack_guard_state_offset() + in_ByteSize(sizeof(StackOverflow::StackGuardState) - 1)),
-           StackOverflow::stack_guard_yellow_reserved_disabled);
+           static_cast<int64_t>(StackOverflow::stack_guard_yellow_reserved_disabled));
 
   __ z_bre(no_reguard);
